Move XPath & XPathElement classes to domUtil instead of application.js
